+++ crates/nostr-relay-pool/src/relay/mod.rs
@@ -64,6 +64,10 @@ pub enum RelayNotification {
         /// Relay Status
         status: RelayStatus,
     },
+    /// Authenticated to relay
+    ///
+    /// <https://github.com/nostr-protocol/nips/blob/master/42.md>
+    Authenticated,
     /// Shutdown
     Shutdown,
 }