+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 * nostr: add NIP-70 support ([Yuki Kishimoto])
 * nostr: add `EventId::LEN` const ([Yuki Kishimoto])
 * nostr: add `UnsignedEvent::ensure_id` method ([Yuki Kishimoto])
+* pool: add `RelayPoolNotification::Authenticated` variant ([Yuki Kishimoto])
 * sdk: add `Client::gift_wrap_to` and `Client::send_private_msg_to` ([reyamir])
 * sdk: add option to autoconnect relay on `Client::add_relay` method call ([Yuki Kishimoto])
 * cli: add `generate` command ([Yuki Kishimoto])